requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
tweepy==4.14.0
linkedin-api==2.0.0
newspaper3k==0.2.8
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0 